+++ rapidhash/src/util/mod.rs
@@ -1,7 +1,5 @@
 //! Utility functions and types for Rapidhash. All should be marked `pub(crate)`.
 
-#[cfg(feature = "std")]
-pub mod chunked_stream_reader;
 pub mod mix;
 pub mod read;
 #[cfg(test)]